Deduplicate knowledge base logic by moving to Thread entity
Extract the duplicated deduplicateKnowledgeBases logic from
get-thread.mapper.ts and tool-assembly.service.ts into a single
getUniqueKnowledgeBases() method on the Thread entity.

This eliminates code duplication across module boundaries and ensures
any future changes to the deduplication logic only need to be made
in one place.

Applied via @cursor push command
