+++ ayunis-core-backend/src/domain/threads/presenters/http/mappers/get-thread.mapper.ts
@@ -4,7 +4,6 @@ import { GetThreadResponseDto } from '../dto/get-thread-response.dto';
 import { MessageDtoMapper } from './message.mapper';
 import { SourceDtoMapper } from './source.mapper';
 import { FindThreadResult } from '../../../application/use-cases/find-thread/find-thread.use-case';
-import type { UUID } from 'crypto';
 
 @Injectable()
 export class GetThreadDtoMapper {
@@ -31,30 +30,11 @@ export class GetThreadDtoMapper {
       updatedAt: thread.updatedAt.toISOString(),
       isAnonymous: thread.isAnonymous,
       isLongChat,
-      knowledgeBases: this.deduplicateKnowledgeBases(thread),
+      knowledgeBases: thread.getUniqueKnowledgeBases(),
     };
   }
 
   toDtoArray(threads: Thread[]): GetThreadResponseDto[] {
     return threads.map((thread) => this.toDto({ thread, isLongChat: false }));
-  }
-
-  /**
-   * Deduplicate knowledge bases by knowledgeBaseId.
-   * The same KB can appear multiple times in thread.knowledgeBaseAssignments
-   * when it was assigned by different origin skills.
-   */
-  private deduplicateKnowledgeBases(
-    thread: Thread,
-  ): { id: UUID; name: string }[] {
-    const seen = new Set<UUID>();
-    const result: { id: UUID; name: string }[] = [];
-    for (const a of thread.knowledgeBaseAssignments ?? []) {
-      if (!seen.has(a.knowledgeBase.id)) {
-        seen.add(a.knowledgeBase.id);
-        result.push({ id: a.knowledgeBase.id, name: a.knowledgeBase.name });
-      }
-    }
-    return result;
   }
 }
